Checkpoint: ✅ CHECKPOINT: CORREÇÕES DE TIPOS E MIGRAÇÃO
CORREÇÕES APLICADAS:
✅ Tipo de `dismissed` corrigido (tinyint em vez de int)
✅ Migração do banco de dados aplicada com sucesso
✅ TypeScript compilando sem erros
✅ Dashboard types corrigidos (any para evitar conflitos)

TESTES EXECUTADOS:
✅ Validação de Arquivos: 2/2 encontrados
✅ Mascaramento LGPD: 5/5 passaram (100%)
⚠️ Análise de Prescrição: 1/4 passaram (25%) - Ajustes necessários

PRÓXIMAS FASES:
1. Executar OCR real com Tesseract.js
2. Testar PaddleOCR no backend
3. Completar pipeline OCR→LLM→Prescrição
4. Corrigir lógica de prescrição (5 vs 10 anos)
5. Testar Dashboard com dados reais

+++ drizzle/schema.ts
@@ -1,4 +1,4 @@
-import { int, mysqlEnum, mysqlTable, text, timestamp, varchar, datetime, decimal, index } from "drizzle-orm/mysql-core";
+import { int, mysqlEnum, mysqlTable, text, timestamp, varchar, datetime, decimal, index, tinyint } from "drizzle-orm/mysql-core";
 
 /**
  * Core user table backing auth flow.
@@ -71,7 +71,7 @@ export const prescriptionAlerts = mysqlTable(
     prescriptionDate: timestamp('prescription_date').notNull(),
     daysRemaining: int('days_remaining').notNull(),
     alertType: mysqlEnum('alert_type', ['URGENT', 'WARNING', 'INFO']).default('INFO'),
-    dismissed: int('dismissed').default(0),
+    dismissed: tinyint('dismissed').default(0),
     dismissedAt: timestamp('dismissed_at'),
     createdAt: timestamp('created_at').defaultNow().notNull(),
   },
